+++ src/screen/dashboard/History/components/export.ts
@@ -1,0 +1,1 @@
+export * from './RecordView';
