+++ src/screen/dashboard/History/export.ts
@@ -1,0 +1,2 @@
+export * from './styles';
+export * from './components/export';
